src/handlers/requirements.txt
-i https://pypi.org/simple
boto3==1.9.238
botocore==1.12.238
certifi==2019.11.28
chardet==3.0.4
docutils==0.15.2
idna==2.9
jmespath==0.9.4
python-dateutil==2.8.1 ; python_version >= '2.7'
pytz==2019.3
requests==2.23.0
s3transfer==0.2.1
six==1.14.0
urllib3==1.25.8 ; python_version >= '3.4'
